+++ cdm_reader_mapper/common/__init__.py
@@ -4,3 +4,4 @@ from __future__ import annotations
 
 from . import json_dict
 from .getting_files import load_file
+from .io_files import get_filename